@@ -3,25 +3,25 @@ import XExpect
 
 @testable import App
 
-final class AudioValidityTests: AppTestCase {
+final class AudioValidityTests: AppTestCase, @unchecked Sendable {
   func testEmptyReaderInvalid() async {
     var audio = Audio.valid
     audio.reader = ""
-    expect(await audio.isValid()).toBeFalse()
+    await expect(audio.isValid()).toBeFalse()
   }
 
   func testM4bLqNotSmallerThanHqInvalid() async {
     var audio = Audio.valid
     audio.m4bSizeHq = 9_000_000
     audio.m4bSizeLq = 9_000_111
-    expect(await audio.isValid()).toBeFalse()
+    await expect(audio.isValid()).toBeFalse()
   }
 
   func testMp3ZipLqNotSmallerThanHqInvalid() async {
     var audio = Audio.valid
     audio.mp3ZipSizeHq = 6_000_000
     audio.mp3ZipSizeLq = 7_000_000
-    expect(await audio.isValid()).toBeFalse()
+    await expect(audio.isValid()).toBeFalse()
   }
 
   func testZeroFilesizesValid() async {
@@ -30,25 +30,25 @@ final class AudioValidityTests: AppTestCase {
     audio.mp3ZipSizeLq = 0
     audio.m4bSizeHq = 0
     audio.m4bSizeLq = 0
-    expect(await audio.isValid()).toBeTrue()
+    await expect(audio.isValid()).toBeTrue()
   }
 
   func testTooSmallNonZeroM4bSizeInvalid() async {
     var audio = Audio.valid
     audio.m4bSizeHq = 1000
-    expect(await audio.isValid()).toBeFalse()
+    await expect(audio.isValid()).toBeFalse()
     audio = Audio.valid
     audio.m4bSizeLq = 1000
-    expect(await audio.isValid()).toBeFalse()
+    await expect(audio.isValid()).toBeFalse()
   }
 
   func testTooSmallNonZeroMp3ZipSizeInvalid() async {
     var audio = Audio.valid
     audio.mp3ZipSizeHq = 1000
-    expect(await audio.isValid()).toBeFalse()
+    await expect(audio.isValid()).toBeFalse()
     audio = Audio.valid
     audio.mp3ZipSizeLq = 1000
-    expect(await audio.isValid()).toBeFalse()
+    await expect(audio.isValid()).toBeFalse()
   }
 
   func testNonSequentialPartsInvalid() async {
@@ -57,6 +57,6 @@ final class AudioValidityTests: AppTestCase {
     part2.audioId = audio.id
     part2.order = 3 // <-- unexpected non-sequential order!!!
     try! await part2.create()
-    expect(await audio.model.isValid()).toBeFalse()
+    await expect(audio.model.isValid()).toBeFalse()
   }
 }
