@@ -3,7 +3,7 @@ import XExpect
 
 @testable import App
 
-final class DocumentTests: AppTestCase {
+final class DocumentTests: AppTestCase, @unchecked Sendable {
   var validDocument: Document {
     var friend = Friend.empty
     friend.lang = .en
@@ -16,28 +16,28 @@ final class DocumentTests: AppTestCase {
   func testEmptyTitleInvalid() async {
     var doc = self.validDocument
     doc.title = ""
-    expect(await doc.isValid()).toBeFalse()
+    await expect(doc.isValid()).toBeFalse()
   }
 
   func testOriginalTitleTooShortInvalid() async {
     var doc = self.validDocument
     doc.originalTitle = "Abc"
-    expect(await doc.isValid()).toBeFalse()
+    await expect(doc.isValid()).toBeFalse()
   }
 
   func testOriginalTitleNotCapitalizedInvalid() async {
     var doc = self.validDocument
     doc.originalTitle = "the life and labors of george dilwynn"
-    expect(await doc.isValid()).toBeFalse()
+    await expect(doc.isValid()).toBeFalse()
   }
 
   func testPublishedDateOutOfBoundsInvalid() async {
     var doc = self.validDocument
     doc.published = 1500
-    expect(await doc.isValid()).toBeFalse()
+    await expect(doc.isValid()).toBeFalse()
     doc = self.validDocument
     doc.published = 1901
-    expect(await doc.isValid()).toBeFalse()
+    await expect(doc.isValid()).toBeFalse()
   }
 
   func testTodosOkForDescriptionsIfEditionsNotLoaded() async {
@@ -45,7 +45,7 @@ final class DocumentTests: AppTestCase {
     doc.description = "TODO"
     doc.partialDescription = "TODO"
     doc.featuredDescription = "TODO"
-    expect(await doc.isValid()).toBeTrue()
+    await expect(doc.isValid()).toBeTrue()
   }
 
   func testTodosInvalidForDescriptionsIfHasNonDraftEdition() async {
@@ -54,31 +54,31 @@ final class DocumentTests: AppTestCase {
       $0.document.partialDescription = "TODO"
       $0.document.featuredDescription = "TODO"
     }.document.model
-    expect(await doc.isValid()).toBeFalse()
+    await expect(doc.isValid()).toBeFalse()
   }
 
   func testNonCapitalizedTitleInvalid() async {
     var doc = self.validDocument
     doc.title = "no Cross, No Crown"
-    expect(await doc.isValid()).toBeFalse()
+    await expect(doc.isValid()).toBeFalse()
   }
 
   func testTooShortTitleInValid() async {
     var doc = self.validDocument
     doc.title = "No"
-    expect(await doc.isValid()).toBeFalse()
+    await expect(doc.isValid()).toBeFalse()
   }
 
   func testNonSluggySlugInvalid() async {
     var doc = self.validDocument
     doc.slug = "This is not A Sluggy Slug"
-    expect(await doc.isValid()).toBeFalse()
+    await expect(doc.isValid()).toBeFalse()
   }
 
   func testMalformedFilenameInvalid() async {
     var doc = self.validDocument
     doc.filename = "This is not A good filename :("
-    expect(await doc.isValid()).toBeFalse()
+    await expect(doc.isValid()).toBeFalse()
   }
 
   func testPrimaryEdition() async {
